requirements.txt
Django==4.2
PyPDF2==3.0.0
Pillow==9.5.0
python-docx==0.8.11
openpyxl==3.1.2
mutagen==1.46.0
opencv-python==4.8.0.74